requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.0.5
mysql-connector-python==8.0.33
python-dotenv==1.0.0
